+++ circuit_maintenance_parser/parsers/zayo.py
@@ -1,14 +1,14 @@
 """Zayo parser."""
 import logging
-import quopri
-from typing import Iterable, Union, Dict
+from typing import Dict
 
 import bs4  # type: ignore
-import dateutil.parser as parser
-from pydantic import ValidationError
+from bs4.element import ResultSet  # type: ignore
 
-from circuit_maintenance_parser.errors import ParsingError, MissingMandatoryFields
-from circuit_maintenance_parser.parser import Html, Impact, CircuitImpact, Maintenance, Status
+import dateutil.parser as parser
+
+from circuit_maintenance_parser.errors import ParsingError
+from circuit_maintenance_parser.parser import Html, Impact, CircuitImpact, Status
 
 # pylint: disable=too-many-nested-blocks,no-member, too-many-branches
 
@@ -25,56 +25,51 @@ class ParserZayo(Html):
     _default_provider = "zayo"
     _default_organizer = "mr@zayo.com"
 
-    def process(self) -> Iterable[Maintenance]:
+    def parse_html(self, soup, data_base):
         """Execute parsing."""
-        data: Dict[str, Union[int, str, Iterable]] = {
-            "provider": self._default_provider,
-            "organizer": self._default_organizer,
-        }
+        data = data_base.copy()
         try:
-            soup = bs4.BeautifulSoup(quopri.decodestring(self.raw), features="lxml")
-            for line in soup.find_all("b"):
-                if isinstance(line, bs4.element.Tag):
-                    if line.text.lower().strip().startswith("maintenance ticket #:"):
-                        data["maintenance_id"] = self.clean_line(line.next_sibling)
-                    elif line.text.lower().strip().startswith("urgency:"):
-                        urgency = self.clean_line(line.next_sibling)
-                        if urgency == "Planned":
-                            data["status"] = Status("CONFIRMED")
-                    elif "activity date" in line.text.lower():
-                        logger.info("Found 'activity date': %s", line.text)
-                        for sibling in line.next_siblings:
-                            text = sibling.text if isinstance(sibling, bs4.element.Tag) else sibling
-                            logger.debug("Checking for GMT date/timestamp in sibling: %s", text)
-                            if "( GMT )" in text:
-                                window = self.clean_line(sibling).strip("( GMT )").split(" to ")
-                                start = parser.parse(window.pop(0))
-                                data["start"] = self.dt2ts(start)
-                                end = parser.parse(window.pop(0))
-                                data["end"] = self.dt2ts(end)
-                                break
-                    elif line.text.lower().strip().startswith("reason for maintenance:"):
-                        data["summary"] = self.clean_line(line.next_sibling)
-                    elif line.text.lower().strip().startswith("date notice sent:"):
-                        stamp = parser.parse(self.clean_line(line.next_sibling))
-                        data["stamp"] = self.dt2ts(stamp)
-                    elif line.text.lower().strip().startswith("customer:"):
-                        data["account"] = self.clean_line(line.next_sibling)
+            self.parse_bs(soup.find_all("b"), data)
+            self.parse_tables(soup.find_all("table"), data)
 
-            data["circuits"] = self.process_circuit_table(soup)
-
-            return [Maintenance(**data)]
-
-        except ValidationError as exc:
-            raise MissingMandatoryFields from exc
+            return [data]
 
         except Exception as exc:
             raise ParsingError from exc
 
-    def process_circuit_table(self, soup: bs4.BeautifulSoup) -> Iterable[CircuitImpact]:
-        """Handles the circuit tables and returns a list of Circuit Impacts."""
+    def parse_bs(self, btags: ResultSet, data: dict):
+        """Parse B tag."""
+        for line in btags:
+            if isinstance(line, bs4.element.Tag):
+                if line.text.lower().strip().startswith("maintenance ticket #:"):
+                    data["maintenance_id"] = self.clean_line(line.next_sibling)
+                elif line.text.lower().strip().startswith("urgency:"):
+                    urgency = self.clean_line(line.next_sibling)
+                    if urgency == "Planned":
+                        data["status"] = Status("CONFIRMED")
+                elif "activity date" in line.text.lower():
+                    logger.info("Found 'activity date': %s", line.text)
+                    for sibling in line.next_siblings:
+                        text = sibling.text if isinstance(sibling, bs4.element.Tag) else sibling
+                        logger.debug("Checking for GMT date/timestamp in sibling: %s", text)
+                        if "( GMT )" in text:
+                            window = self.clean_line(sibling).strip("( GMT )").split(" to ")
+                            start = parser.parse(window.pop(0))
+                            data["start"] = self.dt2ts(start)
+                            end = parser.parse(window.pop(0))
+                            data["end"] = self.dt2ts(end)
+                            break
+                elif line.text.lower().strip().startswith("reason for maintenance:"):
+                    data["summary"] = self.clean_line(line.next_sibling)
+                elif line.text.lower().strip().startswith("date notice sent:"):
+                    stamp = parser.parse(self.clean_line(line.next_sibling))
+                    data["stamp"] = self.dt2ts(stamp)
+                elif line.text.lower().strip().startswith("customer:"):
+                    data["account"] = self.clean_line(line.next_sibling)
+
+    def parse_tables(self, tables: ResultSet, data: Dict):
+        """Parse Table tag."""
         circuits = []
-        tables = soup.find_all("table")
         for table in tables:
             head_row = table.find_all("th")
             if len(head_row) < 5 or [self.clean_line(line) for line in head_row[:5]] != [
@@ -98,4 +93,4 @@ class ParserZayo(Html):
                 if "hard down" in impact.lower():
                     data_circuit["impact"] = Impact("OUTAGE")
                     circuits.append(CircuitImpact(**data_circuit))
-        return circuits
+        data["circuits"] = circuits